#74 refactor: work validator 경로 수정
