+++ src/main/java/com/susanghan_guys/server/work/application/DcaWorkService.java
@@ -10,7 +10,7 @@ import com.susanghan_guys.server.work.dto.request.DcaWorkSubmissionRequest;
 import com.susanghan_guys.server.work.infrastructure.mapper.DcaWorkMapper;
 import com.susanghan_guys.server.work.infrastructure.persistence.WorkRepository;
 import com.susanghan_guys.server.work.infrastructure.saver.WorkSaver;
-import com.susanghan_guys.server.work.validator.DcaWorkValidator;
+import com.susanghan_guys.server.work.application.validator.DcaWorkValidator;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
